cloud_functions/prune_datastore/requirements.txt
functions-framework==3.8.1
google-cloud-datastore==2.20.1
